requirements.txt
flask==2.1.3
pyjwt==2.4.0
requests==2.28.1
gunicorn==20.1.0
Flask-SQLAlchemy